added all time calculations

+++ src/CConflictSolver.h
@@ -11,9 +11,9 @@ public:
 
     static bool sortTimeByStart(const pair<CTime, CTime> &time1, const pair<CTime, CTime> &time2);
 
-    CDatetime getNextFreeDatetime(size_t durationMinutes, const CDatetime &from, size_t ignoreEventId=0, const shared_ptr<CEvent> &newFutureEvent=nullptr);
+    CDatetime getNextFreeDatetime(int durationMinutes, const CDatetime &from, size_t ignoreEventId=0, const shared_ptr<CEvent> &newFutureEvent=nullptr);
 
-    static CTime findFreeTimeInRecurringEvents(vector<pair<CTime, CTime>> &foreverBusyVec, size_t durationMinutes, const CTime &start, const CTime &end) ;
+    static CTime findFreeTimeInRecurringEvents(vector<pair<CTime, CTime>> &foreverBusyVec, int durationMinutes, const CTime &start, const CTime &end) ;
 
 private:
     CCalendar mCalendar;
