+++ src/CEventRecurring.h
@@ -27,7 +27,7 @@ public:
 
     CTime getEndTime() const override;
 
-    [[nodiscard]] size_t getEventDuration() const override;
+    [[nodiscard]] int getEventDuration() const override;
 
     /**
      * Finds out if calendar's event doesn't conflict with the provided event. Since recurring event repeats every
